+++ optparse.d.ts
@@ -1,0 +1,12 @@
+declare module "optparse" {
+  declare class OptionParser {
+    constructor(switches: readonly (readonly [string, string, string])[])
+
+    banner: string
+
+    on(event: string, handler: (opt: string, value: any) => void): void
+    on(unknownOptionhandler: (opt: string, value: any) => void): void
+
+    parse(argv: string[]): void
+  }
+}
